Pull request #48: VLE-459 Enable validation of scenarios with different submodels
UI passes a list of submodel names, but in order to validate them all, we needed to restructure lem_validate_inputfiles. Base scenario data now needs to be validated for each submodel, and zone numbers have to be taken from each scenario.
Handling of free_flow_speedargument is still not handled perfectly, as that is only one variable.

Squashed commit of the following:

commit 0a375e72062d530c4ce47413ee874972c0a21b5d
Author: Jens West <jens.west@traficom.fi>
Date:   Fri Feb 2 13:40:41 2024 +0200

    Fix zone numbers and test

commit 9a2d9107af67808e22961a5bf4424d7ebc0800de
Author: Jens West <jens.west@traficom.fi>
Date:   Fri Feb 2 13:12:30 2024 +0200

    Enable validation of scenarios with different submodels