Pull request #36: VLE-440 Remove addition of one from log-variables with positive parameters
Getting a utility of -Inf is not a problem, so I removed the +1 calculation for variables that are logarithmized and have a positive parameter. For variables with a negative parameter, +1 calculation must be maintained, as it can otherwise cause utility Inf.
As a consequence, we need to be able to handle cases where all alternatives have utilityprobability zero. For instance, the destination choice for walk mode is often restricted to short distances and if the size factor is zero for all the zones near by, all destinations get utility zero. In these cases we get a division by zero, which in the solution in this PR results in zero probabilities.

Squashed commit of the following:

commit 160000c7c98b70ce55d3308c6cf704c9cfb39a9c
Author: Jens West <jens.west@traficom.fi>
Date:   Wed Feb 7 13:46:27 2024 +0200

    Make mode choice fail safe

commit 534db58e7c2b5dd561c5d0444c8685dbdb033001
Author: Jens West <jens.west@traficom.fi>
Date:   Mon Jan 22 11:05:29 2024 +0200

    Fine tune calculation steps

commit e6b6ae606d34defaff2f7c8c54a6c24f48e93358
Author: Jens West <jens.west@traficom.fi>
Date:   Fri Jan 19 16:30:22 2024 +0200

    Remove addition of one from log-variables with positive parameters