Add reading and aggregation of long-distance trips in submodels (#4)
The main feature is the new method ModelSystem.add_external_demand() that is activated for short trip (i.e., submodel) model runs. This method first looks for disaggregated ("koko_suomi") long-distance trip matrices in the base matrix folder and tries to aggregate them. The disaggregated format implies that the long-distance model has been run recently.

If the disaggregated matrices are not found, the next step is to look for aggregated long-distance matrices. These aggregated base matrices should be available for all the submodels, in case the user is not interested in long-distance demand changes. They must fit the zone system of the submodel in question.

This change has some indirect effects on handling of aggregations, which is from now on done both for zone data and for matrices.
--end-assignment-only model runs will assign matrices from the base matrix folder instead of results folder.
Added copying of the input data to result folder, so that the result folder can be used as base input data in another model run.